+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - REST endpoint for editing content entries !3098 @alex.simm
 - Created confirmationDialogue mixin !3140 @AntonBallmaier
 - Removed unused event invitation code #1739 !3149 @AntonBallmaier
+- Move location information in session to user #1769 !3148 @KoeMai
 - Remove unused variables from session #1769 !3161  @KoeMai
 
 ## Dev/Test/CI stuff